+++ changelog.md
@@ -3,3 +3,7 @@
 ## [v1.0.0 (2022-09-10)](https://github.com/Repzo/repzo-quickbooks.git)
 
 ### Added
+
+- [command/item] fix sku & category AND inject tax depend on the QB @maramalshen
+- [command/customer] add formatted_address @maramalshen
+- [command/tax] handle taxCode & taxRate AND add for each tax 2 types additive & inclusive @maramalshen